moving sticky header back into the app and out of presentational components
