--- packages/presentational-components/src/components/sticky-header.md
@@ -1,8 +1,0 @@
-Simple header that sticks to the top of the page.
-
-jsx```
-  <StickyHeader className="sticky">
-    <div>...</div>
-    <div>...</div>
-  </StickyHeader>
-```
